chore: Pass context in to NumericCasts::MarkJsObject()
CreationContext() is going away in V8 11.x. Although there is a
replacement for it, it can be an expensive operation because V8 has to
walk the map to retrieve an object's creation context; so we may as well
try to make it unnecessary.

In MarkJsObject()'s callers, we already had a handle around for the
context, so we can pass it on to MarkJsObject(). This requires replacing
the Isolate* parameter with a Local<Context>. For most of the callers, the
relevant object's creation context is the isolate's current context,
because the object was just created in the isolate.

The only place I'm not sure of is in NativeScriptLongFunctionCallback()
where it seems possible that args.This() might have a different creation
context than the current execution context. However, since we only use one
context in NativeScript this is probably fine, we use the current
execution context right afterwards to set the prototype of the object
anyway.