chore: Remove finalizer callbacks and integrate CPPGC
With V8 11.x no longer supporting finalizer callbacks, we have to replace
our use of them in ObjectManager with CPPGC functionality. That's a
problem, because neither CPPGC nor Java allow exactly replicating the
finalizer callback functionality where you get to 'rescue' an object
that's about to be collected. When integrating two GC systems, either one
side or the other has to allow this, or they have to have some sort of
notification when they reach their last reference.

This commit integrates CPPGC into ObjectManager and simply removes the
finalizer callbacks. JS objects that wrap Java objects are traced by CPPGC
and the liveness of the Java object is checked at trace time so that we
can skip marking the JS object if the Java object is gone.

This is the most conservative approach to replacing finalizer callbacks.
It passes all the tests except for worker tests, but keeps many Java-JS
object pairs alive unnecessarily, causing apps to leak memory. In many
apps this might not be too bad, but it's pretty bad in the test cases that
create thousands of objects in a loop.

Other notable changes:
- The Java runtime makeInstanceWeakAndCheckIfAlive() method is split into
  two new methods, makeInstanceWeak(), and isInstanceAlive().
- JSInstanceInfo gains some functionality and is moved out of
  ObjectManager into its own header/implementation file.
- JavaObjectMap is an abstraction around m_idToObject that integrates
  CPPGC and traces the objects stored in it, rather than allocating a
  Persistent for each one.