WIP - GC callback solution
This uses a GC callback to start and stop counting the number of traces
of a JSInstanceInfo. We assume that if a JSInstanceInfo is traced more
than once, it has a JS object tracing it, and if it's traced only once,
that trace comes from our strong ref in the JavaObjectMap.

At the point where a JSInstanceInfo is only traced by JavaObjectMap, we
can make the strong reference on the Java object into a weak reference.

This collects a lot more objects but is also more unstable, sometimes
crashing in the weakref tests with a Java method call on a null Java
object.